requirements.txt
pandas
numpy
SQLAlchemy
pyodbc
python-dotenv
fpdf
openai
apscheduler
requests
